requirements-python.txt
# Dependências para reconhecimento facial
face-recognition==1.3.0  # Temporariamente comentado devido a problemas de build
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
dlib==19.24.2  # Temporariamente comentado devido a problemas de build
PyMuPDF==1.23.8
